+++ apispec/plugin.py
@@ -12,8 +12,16 @@ class BasePlugin(object):
         :param APISpec spec: APISpec object this plugin instance is attached to
         """
 
-    def definition_helper(self, name, definition, **kwargs):
+    def schema_helper(self, name, definition, **kwargs):
         """Must return definition as a dict."""
+        raise PluginMethodNotImplementedError
+
+    def parameter_helper(self, **kwargs):
+        """Must return parameter component description as a dict."""
+        raise PluginMethodNotImplementedError
+
+    def response_helper(self, **kwargs):
+        """Must return response component description as a dict."""
         raise PluginMethodNotImplementedError
 
     def path_helper(self, path=None, operations=None, **kwargs):