+++ apispec/utils.py
@@ -113,3 +113,17 @@ def dedent(content):
         content = re.sub(re.compile(whitespace_pattern, re.MULTILINE), '', content)
 
     return content.strip()
+
+
+# http://stackoverflow.com/a/8310229
+def deepupdate(original, update):
+    """Recursively update a dict.
+
+    Subdict's won't be overwritten but also updated.
+    """
+    for key, value in original.items():
+        if key not in update:
+            update[key] = value
+        elif isinstance(value, dict):
+            deepupdate(value, update[key])
+    return update