+++ tests/utils.py
@@ -6,5 +6,15 @@ def get_definitions(spec):
         return spec.to_dict()['definitions']
     return spec.to_dict()['components']['schemas']
 
+def get_parameters(spec):
+    if spec.openapi_version.major < 3:
+        return spec.to_dict()['parameters']
+    return spec.to_dict()['components']['parameters']
+
+def get_responses(spec):
+    if spec.openapi_version.major < 3:
+        return spec.to_dict()['responses']
+    return spec.to_dict()['components']['responses']
+
 def get_paths(spec):
     return spec.to_dict()['paths']